Refactor: Improved TokenType, TokenProperties relationship
