+++ api-user/src/main/java/com/user/utils/token/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.user.controller.response;
+package com.user.utils.token;
 
 public record TokenResponse(String accessToken, String refreshToken) {
 }
